+++ src/index.ts
@@ -16,6 +16,8 @@ export {
   MoreThanOneWhereStepError,
   InvalidConditionError,
   DeleteWithoutConditionError,
+  InvalidLimitValueError,
+  InvalidOffsetValueError,
 } from './errors'
 export {
   Database,