+++ test/exported/bitwiseOperator.test.ts
@@ -1,5 +1,5 @@
-import { Builder } from '../src'
-import { database } from './database'
+import { Builder } from '../../src'
+import { database } from '../database'
 
 //Alias
 const table = database.s.public.t.table1