+++ CHANGELOG.md
@@ -10,12 +10,14 @@
 - Functions isEq(), isEq$(), isNe() and isNe$(), those function work with null and non-null values
 - Github Action for unit test at pushing code
 - Codecov badge to README.md
+- Throw error if limit or offer value is NaN or INFINITY
 ### Changed
 - Rename FromStep to SelectFromStep
 - Function selectAsteriskFrom() now can take TableAliases not just Tables
 - Rename `WhereStep` class to `SelectWhereStep`
 ### Breaking Changed
 - Functions eq(), eq$(), ne(), ne$() only deal with non-null values and return equal and not equal operators respectively
+- Rename function "not()" to a getter "not" without parenthesis
 ### Fixed
 - Import Operand from './Operand' instead of './operand'